Refactored UserController
- Changed API path for methods which are depending on AuthToken (/Me)
- Changed location of UserViewModels
